requirements.txt
newsapi-python
beautifulsoup4
google-generativeai
psycopg2-binary
python-dotenv